restructing files, deleting unnecesary variables, folders
