+++ react-app/package.json
@@ -20,6 +20,7 @@
     "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-accessible-accordion": "^5.0.0",
+    "react-axios": "^2.0.6",
     "react-dom": "^18.2.0",
     "react-google-calendar-api": "^2.1.0",
     "react-icons": "^4.6.0",
