+++ react-app/public/index.html
@@ -18,7 +18,10 @@
     />
     <!--- end of the font scripts--->
     <title>Weather app</title>
-    <script src="https://apis.google.com/js/api.js" type="text/javascript"></script>
+    <script
+      src="https://apis.google.com/js/api.js"
+      type="text/javascript"
+    ></script>
   </head>
 
   <body>
