+++ react-app/src/App.css
@@ -28,7 +28,7 @@
   font-size: 23px;
   text-align: center;
   letter-spacing: 0.01em;
-  margin-top: 100px;
+  margin-top: 10px;
   color: #000;
 }
