+++ react-app/src/api-fetches/auto-complete-service.js
