+++ react-app/src/api-fetches/location-converter.js
@@ -1,7 +1,4 @@
-//const address = 'Lange Molenstraat 3, 8200 Sint-Andries, Brugge';
-const addresus =
-  "Basic-Fit Kortrijk Meensestraat, Meensestraat 9, 8500 Kortrijk, België";
-export function ConvertAddress2(address) {
+export function ConvertAddress(address) {
   const API_KEY =
     "RVVfN2ZkYTc5ZTE4NDIzNDNkMTg0NzFlNzI1NjI5N2U5YWY6ZGEwMjlhYjEtOTUwYi00YWRmLWI3MDYtMTZlOGIzZTBjZDcy";
   return new Promise((resolve, reject) => {
