+++ react-app/src/components/Navbar.jsx
@@ -8,7 +8,7 @@ import { Layout } from "antd";
 import { useNavigate } from "react-router-dom";
 import ProfilePicture from "../components/ProfilePicture";
 
-const Navbar = ({setReveal}) => {
+const Navbar = ({ setReveal }) => {
   const { Header } = Layout;
   const [navbarOpen, setNavbarOpen] = useState(false);
   const navigate = useNavigate();
@@ -49,7 +49,7 @@ const Navbar = ({setReveal}) => {
           <div className="forms-container">
             <Layout style={{ backgroundColor: "transparent" }}>
               <Header className="form-buttons">
-                <Appheader setReveal={setReveal}/>
+                <Appheader setReveal={setReveal} />
               </Header>
               <NavigationRoutes />
             </Layout>
